--- packages/transaction-pool-server/source/controllers/transaction-pool.ts
@@ -1,26 +1,0 @@
-import Hapi from "@hapi/hapi";
-import { AbstractController } from "@mainsail/api-common";
-import { inject, injectable } from "@mainsail/container";
-import { Contracts, Identifiers } from "@mainsail/contracts";
-
-@injectable()
-export class TransactionsController extends AbstractController {
-	@inject(Identifiers.TransactionPool.Processor)
-	private readonly processor!: Contracts.TransactionPool.Processor;
-
-	public async store(request: Hapi.Request) {
-		const result = await this.processor.process(
-			// @ts-ignore
-			request.payload.transactions.map((transaction: string) => Buffer.from(transaction, "hex")),
-		);
-		return {
-			data: {
-				accept: result.accept,
-				broadcast: result.broadcast,
-				excess: result.excess,
-				invalid: result.invalid,
-			},
-			errors: result.errors,
-		};
-	}
-}
